Move Partner to enums
