+++ modules/testkit/src/main/scala/lucuma/core/model/arb/ArbRole.scala
@@ -3,8 +3,8 @@
 
 package lucuma.core.model.arb
 
+import lucuma.core.enums.Partner
 import lucuma.core.model.GuestRole
-import lucuma.core.model.Partner
 import lucuma.core.model.Role
 import lucuma.core.model.ServiceRole
 import lucuma.core.model.StandardRole
